Updates for Snellius & train-val-test split
I made some changes for the purpose of not having to load rdflib, because I want to avoid using a .venv in Snellius and just want use the available modules (which has everything I need except rdflib). Now, I prepare the OfficeGraph data to a torch format in my local computer.

Also updated the train/val/test split, trying to account for seasonality and preventing leakage.