Update to work with polygon data
Now, the OfficeGraphBuilder class can work with either polygon or CSV floorplan input. I do not intend to develop the CSV floor plans further (i.e., build files for different floors), but I wanted to leave it there since it is already done and works with the model. So, a sanity check basically further down the line. From now on, I intend to work with just polygon room data that I build using InkScape. So, I will optimize the code mostly for that.