Bug fix for LightGBMPruningCallback parameters
Pass 'metric_name' and 'valid_name' explicitly to LightGBMPruningCallback for improved clarity and compatibility. This change ensures the callback monitors the correct metric and validation set.

I was getting the following error:
```
ValueError: The entry associated with the validation name "valid_0" and the metric name "mae" is not found in the evaluation result list [('train', 'l1', 0.09622110391301757, False), ('valid', 'l1', 0.1013087>
```